Move sequence type to individual file
