+++ src/shared/types/sequence.ts
@@ -1,0 +1,7 @@
+export type Sequence = {
+  value: string;
+  length: number;
+  molWeight: number;
+  crc64: string;
+  md5: string;
+};
